Add files via upload
This repository is part of my internship with CODSOFT. It contains the first task assigned to me, which is a Java-based game program titled NumberGame.java. The program demonstrates basic game development principles using Java.
NumberGame.java is a console-based game where the player is prompted to guess a number within a specified range. The game provides feedback on whether the player's guess is too high or too low and tracks the number of attempts.
Technologies Used : Java